Version 7 Update 1/17/2024
Updates:
- Modified the way the announcer is triggered when the player picks up items and weapons. I changed the Pickup Flash actor and modified the attenuation of the "Full" sound.

While testing a multiplayer game using bots, I was able to hear exactly what the bot was picking up even when it was far away from my position. I altered the way the player would hear the announcer so the only time you'll hear him is if you are picking something up and not another player.