Update 12-7-2024
Updated and fixed some things. I added in the narrator for the menu items pertaining to the Item Checker, Coordinate Checker, Map Checker, and Proximity Detector.
I, too, modified the sound for the Proximity Detector; I put in some White Noise. There ,too, is an alternative sound which is a more audible wind sound incase if you'd want to go with this option.
Next, I changed the default footstep sound to the footsteps from Unreal - I feel they're less annoying.
Lastly, I updated some things in the sound bindings scripts and fixed a tiny typo in the ACS script for the powerup wear-off system.