Quick update
As per notice regarding the door/switch beacons and how they can be killed via "Kill Monsters" command, I took the liberty in cleaning up the beacon scripts using inheritance and removing old, unused stuff. Credit to Proydoha for encouraging me to work on this and inspiring the idea :)

Other scripts have been modified. They'll be an update for the Decorations - worked on the explosive barrel.